src/requirements.txt
tensorflow==2.1
tf2cv
numpy
matplotlib
pillow
h5py<3.0.0